feat(friendships): auto-create contact when accepting friend request
