+++ src/features/friendships/domain/repositories/friendship-repository.ts
@@ -13,4 +13,5 @@ export interface FriendshipRepository {
   listFriendships: () => Promise<Friendship[]>
   findUserByPhone: (phone: string) => Promise<string | null>
   findUsersByPhones: (phones: string[]) => Promise<Array<{ phone: string, userId: string }>>
+  findPhonesByUserIds: (userIds: string[]) => Promise<Array<{ userId: string, phone: string }>>
 }
