+++ src/locales/en.json
@@ -347,6 +347,7 @@
     "requestTo": "Request to",
     "verifyPhoneHint": "Verify your phone to connect with other users.",
     "friendsListLink": "Friend requests",
+    "contactCreated": "Contact created for new friend",
     "verificationNotice": {
       "title": "Phone discoverability notice",
       "body": "Once verified, other verified TouringBuddy users who add this number to their contacts will know an account exists for it and can send a friend request. Your identity is not revealed beyond this. Unverified phones are never discoverable. You can deny any incoming request without justification and without penalty. You may revoke friendship connections in the future.",
